% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine_functions.R
\name{affine_exp}
\alias{affine_exp}
\title{affine_exp}
\usage{
affine_exp(P, W, chol = FALSE, P.sq = NULL, ...)
}
\arguments{
\item{P}{Matrix: The base point on the manifold}

\item{W}{Matrix: If applying the Exp map, W is a tangent vector at P.  If applying the Log map, W is another point on the manifold.}

\item{chol}{Logical: Should Cholesky square root be used?}

\item{P.sq}{Matrix: Optional. The square root of P, to help speed calculation.}

\item{...}{Optional arguments to be passed on to affine_map}
}
\value{
the Riemannian exponential map under the affine-invariant geometry
}
\description{
Calculates the Riemannian exponential map under the affine-invariant geometry
}
\examples{

TBA
}
