% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine_functions.R
\name{affine_inner_product}
\alias{affine_inner_product}
\title{affine_inner_product}
\usage{
affine_inner_product(P, U, V, P.sq = NULL)
}
\arguments{
\item{P}{Matrix - the base point on the manifold.  Must be symmetric positive definite.}

\item{U}{Matrix - A symmetric matrix the same dimension as P}

\item{V}{Matrix - A symmetric matrix the same dimension as P}

\item{P.sq}{Matrix - If the square root of P has been calculated, you can input it for use here.}
}
\value{
A scalar inner product value
}
\description{
affine_inner_product
}
\details{
This function calculates the affine invariant inner product value at P as

\if{html}{\out{<div class="sourceCode">}}\preformatted{ <U, V>_P = tr(U P^\{-1\} V P^\{-1\})
}\if{html}{\out{</div>}}
}
\examples{

set.seed(2021)

P <- diag(1, 3)
A <- matrix(rnorm(9), 3, 3)
A <- (A + t(A))/2
B <- matrix(rnorm(9), 3, 3)
B <- (B + t(B))/2

affine_inner_product(P = P, U = A, V = B)

}
