% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine_functions.R
\name{affine_norm}
\alias{affine_norm}
\title{affine_norm}
\usage{
affine_norm(P, W, P.sq = NULL)
}
\arguments{
\item{P}{Matrix: the base point in the manifold}

\item{W}{Matrix: The tangent vector at P you wish to calculate the norm of.}

\item{P.sq}{Matrix: the square root of P, to speed calculations}
}
\value{
A number value of the norm of W at P.
}
\description{
This will calculate the norm of a vector W in the tangent space of P.  This is used directly to
calculate the Frechet mean.
}
\details{
This performs the calculation

||W||_P = ||P^{-1/2} W P^{-1/2}||_I

Which is calculated as the Frobenius norm of P^{-1/2} W P^{-1/2}
}
\examples{

## We will push a vector up to the tangent space of P with the logarithm map
## and then calculate the norm.

P <- cobre$cors[[1]]
W <- cobre$cors[[2]]

## Push W to a tangent vector using the Logarithm

W_vector <- affine_map(P, W, method = "log")

## Now get the norm of W_vector

affine_norm(P, W_vector)

## Note, this method of pushing W up, and then getting the norm, is equivalent to calculating
## the affine invariant distance

affine_dist(slice(cobre, 1:2), "cors")

}
