% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_pls_affine.R
\name{calculate_c_affine}
\alias{calculate_c_affine}
\title{Calculate the weights C, affine variant}
\usage{
calculate_c_affine(
  mu,
  C,
  Y,
  t,
  max.iter = 50,
  tol = 1e-06,
  scale_c = TRUE,
  ...
)
}
\arguments{
\item{mu}{Matrix: mean of Y}

\item{C}{Matrix: Current estimate of C}

\item{Y}{List: List of data matrices Y. If Y is a data frame, please use var.name.Y}

\item{t}{Numeric: Numeric vector of current t scores}

\item{max.iter}{Numeric: Maximum number of iterations}

\item{tol}{Numeric: Tolerance to work at.}

\item{...}{Optional arguments
\itemize{
\item mc.cores: Numeric, number of cores to run on
\item check: Logical, Should we check our inputs
\item verbose: Logical, Do we want to print progress output?
\item var.name.Y: Character, name of column of data if Y is data frame.
\item mu.sq: Matrix, square root of mu to speed computation
\item tau: Numeric, initial step size.
}}
}
\value{
The new Weights C
}
\description{
A wrapper for easy calculation of C.
}
\examples{
TBA
}
