% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_pls_euclidean.R
\name{calculate_c_euclidean}
\alias{calculate_c_euclidean}
\title{Calculate Y-weights, Euclidean}
\usage{
calculate_c_euclidean(Y, t, ...)
}
\arguments{
\item{Y}{Matrix: matrix of data values of size n x q}

\item{t}{Numeric: Scores for X of size n x 1}

\item{...}{}
}
\value{
Numeric vector of size q X 1, it is a direction in Y-space
}
\description{
Calculate Y-weights, Euclidean
}
\examples{
TBA
}
