% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_pls_affine.R
\name{calculate_direction_affine}
\alias{calculate_direction_affine}
\title{Calculate the direction vector, affine invariant}
\usage{
calculate_direction_affine(mu, P, X, t, max.iter = 50, tol = 1e-06, ...)
}
\arguments{
\item{mu}{Matrix: Mean of the X}

\item{P}{Matrix: Current estimate of direction vector}

\item{X}{List: List of data matrices X. If as data frame, provide optional argument var.name.X}

\item{t}{Numeric: Vector of current scores.}

\item{max.iter}{Numeric: Maximum number of iterations.}

\item{tol}{Numeric: Tolerance to work at.}

\item{...}{Optional arguments
\itemize{
\item var.name.X: Character, if X is data frame, name of column with data.
\item mu.sq: Matrix: Square root of mu to speed up computation.
\item tau: Numeric, initial step size.
\item mc.cores: Numeric, if using parallel, number of cores.
\item check: Logical, should we check our inputs.
\item verbose: Logical, should we display progress output.
}}
}
\value{
A new, updated direction
}
\description{
Using gradient descent, calculate the direction vector in the affine invariant geometry.
}
\examples{
TBA

}
