% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_pls_euclidean.R
\name{calculate_direction_euclidean}
\alias{calculate_direction_euclidean}
\title{Calculate the direction of interest in Euclidean NIPALS}
\usage{
calculate_direction_euclidean(X, u, ...)
}
\arguments{
\item{X}{Matrix: matrix of data values of size n x p}

\item{u}{Numeric: Current estimate of scores for Y}

\item{...}{}
}
\value{
Numeric vector of size p X 1, it is a direction in X-space
}
\description{
Calculate the direction of interest in Euclidean NIPALS
}
\examples{
TBA
}
