% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_pls.R
\name{calculate_loading_X}
\alias{calculate_loading_X}
\title{Calculate the X-loadings in Riemannian PLS}
\usage{
calculate_loading_X(
  X,
  t,
  muX = NULL,
  max.iter = 50,
  tol = 1e-06,
  tau = 1/2,
  P = NULL,
  ...
)
}
\arguments{
\item{X}{List or matrix: Predictor data.  List for affine, matrix for euclidean}

\item{t}{Numeric: Vector of scores for the X-data.}

\item{muX}{Matrix:  If X is affine, the frechet mean of X}

\item{max.iter}{Numeric: The maximum number of iterations in the while loops}

\item{tol}{Numeric: The tolerance/precision level to work at}

\item{tau}{Numeric: Step size in the gradient descent.}

\item{P}{Matrix: Initial estimate of the loadings P.}

\item{...}{Optional arguments.
- type_x: String. The data type.  Can be "euclidean" or "affine".  Defualts to affine.
- muX.sq: Matrix.  The square roots of muX.
- other arguments passed to calculate_p.}
}
\value{
The loading vector P
}
\description{
Calculate the X-loadings in Riemannian PLS
}
\examples{
TBA
}
