% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_pls_affine.R
\name{calculate_p_affine}
\alias{calculate_p_affine}
\title{Calculate the loadings P, affine variant}
\usage{
calculate_p_affine(mu, P, X, t, max.iter = 50, tol = 1e-06, j = 0, ...)
}
\arguments{
\item{mu}{Matrix: mean of X}

\item{P}{Matrix: Current estimate of P}

\item{X}{List: List of data matrices X. If X is a data frame, please use var.name.X}

\item{t}{Numeric: Numeric vector of current t scores}

\item{max.iter}{Numeric: Maximum number of iterations}

\item{tol}{Numeric: Tolerance to work at.}

\item{...}{Optional arguments
\itemize{
\item mc.cores: Numeric, number of cores to run on
\item check: Logical, Should we check our inputs
\item verbose: Logical, Do we want to print progress output?
\item var.name.X: Character, name of column of data if X is data frame.
\item mu.sq: Matrix, square root of mu to speed computation
\item tau: Numeric, initial step size.
}}
}
\value{
The new Loadings P.
}
\description{
A wrapper for easy calculation of P.
}
\examples{
TBA
}
