% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_pls_euclidean.R
\name{calculate_p_euclidean}
\alias{calculate_p_euclidean}
\title{Calculate X-loadings, Euclidean}
\usage{
calculate_p_euclidean(X, t, ...)
}
\arguments{
\item{X}{Matrix: matrix of data values of size n x p}

\item{t}{Numeric: Scores for X of size n x 1}

\item{...}{}
}
\value{
Numeric vector of size p X 1, it is a direction in X-space
}
\description{
Calculate X-loadings, Euclidean
}
\examples{
TBA
}
