% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_pls_affine.R
\name{calculate_parallel_transport}
\alias{calculate_parallel_transport}
\title{Parallel transport in the affine-invariant geometry}
\usage{
calculate_parallel_transport(P, Q, W, method = c("symmetric", "non-symmetric"))
}
\arguments{
\item{P}{Matrix: base point matrix}

\item{Q}{Matrix: Endpoint matrix}

\item{W}{Matrix: Vector in T_P M}

\item{method}{Character: Either symmetric or non-symmetric.  Changes the matrix used to calculated the parallel transport.}
}
\value{
A vector in T_Q M
}
\description{
Calculate the affine invariant parallel transport of W from T_PM to T_Q M
}
\examples{
TBA
}
