% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_pls_affine.R
\name{calculate_q_affine}
\alias{calculate_q_affine}
\title{Calculate the loadings Q, affine variant}
\usage{
calculate_q_affine(mu, Q, Y, u, max.iter = 50, tol = 1e-06, j = 0, ...)
}
\arguments{
\item{mu}{Matrix: mean of Y}

\item{Q}{Matrix: Current estimate of Q}

\item{Y}{List: List of data matrices Y. If Y is a data frame, please use var.name.Y}

\item{u}{Numeric: Numeric vector of current u scores}

\item{max.iter}{Numeric: Maximum number of iterations}

\item{tol}{Numeric: Tolerance to work at.}

\item{...}{Optional arguments
\itemize{
\item mc.cores: Numeric, number of cores to run on
\item check: Logical, Should we check our inputs
\item verbose: Logical, Do we want to print progress output?
\item var.name.Y: Character, name of column of data if Y is data frame.
\item mu.sq: Matrix, square root of mu to speed computation
\item tau: Numeric, initial step size.
}}
}
\value{
The new Loadings Q.
}
\description{
A wrapper for easy calculation of Q.
}
\examples{
TBA
}
