% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_pls_euclidean.R
\name{calculate_q_euclidean}
\alias{calculate_q_euclidean}
\title{Calculate Y-loadings, Euclidean}
\usage{
calculate_q_euclidean(Y, u, ...)
}
\arguments{
\item{Y}{Matrix: matrix of data values of size n x q}

\item{u}{Numeric: Scores for U of size n x 1}

\item{...}{}
}
\value{
Numeric vector of size q X 1, it is a direction in X-space
}
\description{
Calculate Y-loadings, Euclidean
}
\examples{
TBA
}
