% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_pls.R
\name{calculate_regression}
\alias{calculate_regression}
\title{Calculate the inner relationship for Riemannian PLS}
\usage{
calculate_regression(t, u)
}
\arguments{
\item{t}{Numeric: Vector of scores for X}

\item{u}{Numeric: Vector of scores for Y}
}
\value{
A list containing the linear model u ~ t, as well as the coefficients.
}
\description{
Calculate the inner relationship for Riemannian PLS
}
\examples{
TBA
}
