% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_pls_affine.R
\name{calculate_score_affine}
\alias{calculate_score_affine}
\title{Calculate the score for a subject, affine variant}
\usage{
calculate_score_affine(
  t,
  W,
  X,
  muX,
  tau = 1/2,
  max.iter = 50,
  tol = 1e-06,
  ...
)
}
\arguments{
\item{t}{Numeric: Initial estimate of t.}

\item{W}{Matrix: Current estimate of weight direction W.}

\item{X}{Matrix: Data point X.}

\item{muX}{Matrix: The mean of the X.}

\item{tau}{Numeric: Initial step size.}

\item{max.iter}{Numeric: Maximum number of iterations for while loop.}

\item{tol}{Numeric: Desired tolerance.}

\item{...}{\itemize{
\item check: Logical, should we check our variables.
}}
}
\value{
A new estimate of the score.
}
\description{
Using gradient descent, calculate the score for a single given subject.
}
\examples{
TBA
}
