% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_pls_affine.R
\name{calculate_score_affine_full}
\alias{calculate_score_affine_full}
\title{Calculate all scores, affine variant}
\usage{
calculate_score_affine_full(
  mu,
  W,
  X,
  t,
  tau = 1/2,
  max.iter = 50,
  tol = 1e-06,
  ...
)
}
\arguments{
\item{mu}{Matrix: Mean of the X.}

\item{W}{Matrix: Current estimate of direction W.}

\item{X}{List: List of the data matrices X.}

\item{t}{Numeric: Vector of current estimates of t.}

\item{tau}{Numeric: Initial Step size}

\item{max.iter}{Numeric: Maximum number of iterations}

\item{tol}{Numeric: Tolerance to work at.}

\item{...}{Optional arguments
\itemize{
\item check: Logical, should we check our estimates
\item mc.cores: Numeric, if parallelising.
}}
}
\value{
An updated vector of scores.
}
\description{
Calculate all scores, affine variant
}
\examples{
TBA
}
