% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_pls_euclidean.R
\name{calculate_score_euclidean}
\alias{calculate_score_euclidean}
\title{Calculate scores in Euclidean NIPALS}
\usage{
calculate_score_euclidean(X, W, ...)
}
\arguments{
\item{X}{Matrix: matrix of data values of size n x p}

\item{W}{Numeric: Vector of weights of size p x 1}

\item{...}{}
}
\value{
Numeric vector of length n containing the scores for X in the direction W.
}
\description{
Calculate scores in Euclidean NIPALS
}
\examples{
TBA
}
