% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_pls_affine.R
\name{calculate_t_affine}
\alias{calculate_t_affine}
\title{Calculate the score t, affine variant}
\usage{
calculate_t_affine(mu, W, X, t, tau = 1/2, max.iter = 50, tol = 1e-06, ...)
}
\arguments{
\item{mu}{Matrix: mean of X}

\item{W}{Matrix: Current estimate of W}

\item{X}{List: List of data matrices X}

\item{t}{Numeric: Numeric vector of current t scores}

\item{tau}{Numeric: Initial step size.}

\item{max.iter}{Numeric: Maximum number of iterations}

\item{tol}{Numeric: Tolerance to work at.}

\item{...}{Optional arguments
\itemize{
\item mc.cores: Numeric, number of cores to run on
\item check: Logical, Should we check our inputs
\item verbose: Logical, Do we want to print progress output?
}}
}
\value{
The new scores t
}
\description{
A wrapper for easy calculation of t.
}
\examples{
TBA
}
