% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_pls_euclidean.R
\name{calculate_t_euclidean}
\alias{calculate_t_euclidean}
\title{Calculate X-scores, Euclidean}
\usage{
calculate_t_euclidean(X, W, ...)
}
\arguments{
\item{X}{Matrix: matrix of data values of size n x p}

\item{W}{Numeric: Vector of weights of size p x 1}

\item{...}{}
}
\value{
Numeric: Vector of weights of size p x 1
}
\description{
Calculate X-scores, Euclidean
}
\examples{
TBA
}
