% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_pls_affine.R
\name{calculate_u_affine}
\alias{calculate_u_affine}
\title{Calculate the score u, affine variant}
\usage{
calculate_u_affine(mu, C, Y, u, tau, max.iter = 50, tol = 1e-06, ...)
}
\arguments{
\item{mu}{Matrix: mean of X}

\item{C}{Matrix: Current estimate of C}

\item{Y}{List: List of data matrices Y}

\item{u}{Numeric: Numeric vector of current u scores}

\item{tau}{Numeric: Initial step size.}

\item{max.iter}{Numeric: Maximum number of iterations}

\item{tol}{Numeric: Tolerance to work at.}

\item{...}{Optional arguments
\itemize{
\item mc.cores: Numeric, number of cores to run on
\item check: Logical, Should we check our inputs
\item verbose: Logical, Do we want to print progress output?
}}
}
\value{
The new scores u
}
\description{
A wrapper for easy calculation of u.
}
\examples{
TBA
}
