% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_pls_euclidean.R
\name{calculate_u_euclidean}
\alias{calculate_u_euclidean}
\title{Calculate Y-scores, Euclidean}
\usage{
calculate_u_euclidean(Y, C, ...)
}
\arguments{
\item{Y}{Matrix: matrix of data values of size n x q}

\item{C}{Numeric: Vector of weights of size q x 1}

\item{...}{}
}
\value{
Numeric: Vector of weights of size q x 1
}
\description{
Calculate Y-scores, Euclidean
}
\examples{
TBA
}
