% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_pls_euclidean.R
\name{calculate_w_euclidean}
\alias{calculate_w_euclidean}
\title{Calculate X-weights, Euclidean}
\usage{
calculate_w_euclidean(X, u, ...)
}
\arguments{
\item{X}{Matrix: matrix of data values of size n x p}

\item{u}{Numeric: Scores for Y of size n x 1}

\item{...}{}
}
\value{
Numeric vector of size p X 1, it is a direction in X-space
}
\description{
Calculate X-weights, Euclidean
}
\examples{
TBA
}
