% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_right_input.R
\name{check_input_values}
\alias{check_input_values}
\title{check_input_values}
\usage{
check_input_values(
  t,
  u,
  W,
  C,
  X,
  Y,
  muX,
  muY,
  lambda_t,
  lambda_u,
  lambda_1,
  lambda_2,
  nu,
  tau
)
}
\arguments{
\item{tau}{}
}
\description{
This function throws an error if any of the parameters in Riemannian PLS
are not defined as they should be.  This is not for external use.
}
