% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_right_input.R
\name{check_right_input}
\alias{check_right_input}
\title{check_right_input}
\usage{
check_right_input(df, var.name = NULL)
}
\arguments{
\item{df}{\itemize{
\item data the user input, should be a list, array, or dataframe
}}

\item{var.name}{\itemize{
\item variable name when df is a dataframe
}}
}
\value{
the data as a list
}
\description{
check_right_input
}
\details{
Not for external use.  This function will convert the user input into a list.
It will also extract the number of ROI and the number of subject
}
\examples{
df <- check_right_input(cobre, "cors")
}
