% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_right_input.R
\name{check_right_input_loadings}
\alias{check_right_input_loadings}
\title{check_right_input_loadings}
\usage{
check_right_input_loadings(X, muX, P, t, ...)
}
\arguments{
\item{...}{}
}
\description{
Check that the loading inputs are correct
}
\examples{

TBA
}
