% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclid_dist.R
\name{corr_dist}
\alias{corr_dist}
\title{corr_dist}
\usage{
corr_dist(df, var.name = NULL, sub_ids = NULL)
}
\arguments{
\item{df}{Subject array/list/df.}

\item{var.name}{Variable name containing correlations, df only, must be string.}

\item{sub_ids}{A character list of subject id's to keep track of.  Defaults to NULL.}
}
\value{
A dist object containing the Correlation distances.
}
\description{
This is a function to calculate the Euclidean distance between correlation matrices.  Dependent
packages are dplyr, purrr.  See details for more information.
}
\details{
This will take a list of correlation matrices, take the upper triangle of each,
and calculates the correlation distance between subjects based on this vector as defined by

d(x, y) = \sqrt{2(1 - \rho(x, y))}

where \rho(x, y) is the correlation between vectors x and y.
}
