% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_pls.R
\name{deflate_X}
\alias{deflate_X}
\title{Deflate the X-data in Riemannian PLS}
\usage{
deflate_X(X, t, P, muX = NULL, ...)
}
\arguments{
\item{X}{List or matrix: Predictor data.  List for affine, matrix for euclidean}

\item{t}{Numeric: Vector of scores for the X-data.}

\item{P}{Matrix: The loading matrix for the X-data and scores t.}

\item{muX}{Matrix:  If X is affine, the frechet mean of X}

\item{...}{Optional list
- type_x: String. The data type.  Can be "euclidean" or "affine".  Defualts to affine.
- muX.sq: Matrix.  The square roots of muX.
- other arguments passed to deflate}
}
\value{
A deflated version of the X-data, removing the information from the previous PLS step.
}
\description{
Deflate the X-data in Riemannian PLS
}
\examples{
TBA
}
