% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_pls_affine.R
\name{deflate_X_affine}
\alias{deflate_X_affine}
\title{Deflate X, affine-variant}
\usage{
deflate_X_affine(X, P, t, muX, par = TRUE, ...)
}
\arguments{
\item{X}{List: List of PSD matrices, the X data.}

\item{P}{Matrix: The loadings from the PLS model.}

\item{t}{Numeric: Vector of estimated scores from the PLS model.}

\item{muX}{Matrix: Frechet mean of the X}

\item{par}{Logical: Use parallel transport or more direct group action?}

\item{...}{Optional arguments
- mu.sq: Square root of the matrix muX to speed up computation.}
}
\value{
A list the same size as X containing deflated data
}
\description{
Deflate the X-data in the affine-PLS model.
}
\examples{
TBA
}
