% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_pls_euclidean.R
\name{deflate_X_euclidean}
\alias{deflate_X_euclidean}
\title{Deflate the X-values in Euclidean space}
\usage{
deflate_X_euclidean(X, t, P, ...)
}
\arguments{
\item{X}{Matrix: matrix of data values of size n x p}

\item{t}{Numeric: Scores for X of size n x 1}

\item{P}{Numeric: Vector of loadings of size p x 1}

\item{...}{}
}
\value{
A new set of X-values that have removed the information from the previous direction
}
\description{
Deflate the X-values in Euclidean space
}
\examples{
TBA
}
