% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_pls.R
\name{deflate_Y}
\alias{deflate_Y}
\title{Deflate the Y-data in Riemannian PLS}
\usage{
deflate_Y(Y, C, Q, B, muY = NULL, ...)
}
\arguments{
\item{Y}{List or matrix: Response data.  List for affine, matrix for euclidean}

\item{C}{Matrix: The weights matrix for the Y-data (may need to change to Y-loadings)}

\item{Q}{Matrix: The loadings matrix for the Y-data (may need to change to Y-loadings)}

\item{B}{Model: The inner relation model of u ~ t}

\item{muY}{Matrix:  If Y is affine, the frechet mean of Y}

\item{...}{Optional list
- type_y: String. The data type.  Can be "euclidean" or "affine".  Defualts to affine.
- muY.sq: Matrix.  The square roots of muY.
- other arguments passed to deflate}
}
\value{
A deflated version of the Y-data, removing the information from the previous PLS step.
}
\description{
Deflate the Y-data in Riemannian PLS
}
\examples{
TBA
}
