% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_pls_affine.R
\name{deflate_Y_affine}
\alias{deflate_Y_affine}
\title{Deflate Y, affine variant}
\usage{
deflate_Y_affine(Y, Q, B, muY, par = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
Deflate the response data Y in the affine-PLS model
}
