% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_pls_euclidean.R
\name{deflate_Y_euclidean}
\alias{deflate_Y_euclidean}
\title{Deflate the Y-values in Euclidean space}
\usage{
deflate_Y_euclidean(Y, C, B, ...)
}
\arguments{
\item{Y}{Matrix: matrix of data values of size n x q}

\item{C}{Numeric: Vector of weights of size q x 1}

\item{B}{Linear model: A linear model fitting the Y-scores to the X-scores.  The inner relation}

\item{...}{}
}
\value{
A new set of Y-values that have removed the information from the previous direction
}
\description{
Deflate the Y-values in Euclidean space
}
\examples{
TBA
}
