% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_short_functions.R
\name{double_centre}
\alias{double_centre}
\title{double_centre}
\usage{
double_centre(X)
}
\arguments{
\item{X}{Matrix: A square, numeric matrix.}
}
\value{
A double-centred version of X.
}
\description{
This function will double centre the input matrix, that is, it will make the row and column means 0.
}
\details{
Returns the matrix J X J where J is the centring matrix

J = I - 1/n 1 t(1).
}
\examples{

set.seed(2022)
X <- matrix(rnorm(9), 3, 3)

centred_X <- double_centre(X)

all.equal(apply(centred_X, 1, mean), rep(0, 3))
all.equal(apply(centred_X, 2, mean), rep(0, 3))

}
