% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclid_dist.R
\name{euclid_dist}
\alias{euclid_dist}
\title{euclid_dist}
\usage{
euclid_dist(df, var.name = NULL, sub_ids = NULL)
}
\arguments{
\item{df}{Subject array/list/df.}

\item{var.name}{Variable name containing correlations, df only, must be string.}

\item{regularise}{Do we want to regularise the matrices.  Default is FALSE.}

\item{lambda}{Regularisation value.}
}
\value{
A dist object containing the Euclidean distances.
}
\description{
This is a function to calculate the Euclidean distance between correlation matrices.  Dependent
packages are dplyr, purrr.  See details for more information.
}
\details{
This will take a list of correlation matrices, take the upper triangle of each,
and calculates the Euclidean distance between subjects based on this vector.
}
