% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine_differentials.R
\name{exponential_adjoint}
\alias{exponential_adjoint}
\title{exponential_adjoint}
\usage{
exponential_adjoint(A, V, t = 1, tol = 1e-08, ...)
}
\arguments{
\item{A}{Matrix: Evaluation point.  Must be symmetric.}

\item{V}{Matrix: Directional point.  Must be symmetric}

\item{t}{Numeric: Time to evaluate at, i.e. exp(tA)}

\item{tol}{Numeric: Tolerance at which we determine if eigenvalues are "close"}

\item{...}{Optional arguments
-  method : The method to pass to expm or logm.  Only accepts Eigen or Higham08}
}
\value{
A matrix that is the adjoint of d_A Exp applied to V.
}
\description{
Calculate the adjoint of the derivative of the matrix exponential map
}
\details{
Calculates the adjoint of d_A Exp as

(d_A Exp)^\dagger (V) = Exp(A)^{-1/2} (d_A Exp(Exp(A^{-1/2}) V Exp(A)^{-1/2})) Exp(A^{1/2}).

This has the property that

< d_AExp(W), V >_{Exp(A)} = < W, (d_A Exp)^\dagger(V) >_I

For all symmetric matrices W and V.
}
