% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine_differentials.R
\name{exponential_adjoint_general}
\alias{exponential_adjoint_general}
\title{exponential_adjoint_general}
\usage{
exponential_adjoint_general(P, A, V, t = 1, P.sq = NULL, tol = 1e-08, ...)
}
\arguments{
\item{P}{Matrix: Base point.  Must be symmetric positive definite.}

\item{A}{Matrix: Evaluation point.  Must be symmetric.}

\item{V}{Matrix: Directional point.  Must be symmetric.}

\item{t}{Numeric: Time to evaluate at, i.e. exp(tA)}

\item{P.sq}{Matrix: Optional input of swaure root of P to speed up calculations.}

\item{tol}{Numeric: Tolerance at which we determine if eigenvalues are "close"}

\item{...}{Optional argument to be passed to exponential_adjoint}
}
\value{
A matrix that is the adjoint of d_A Exp_P applied to V.
}
\description{
Applies the chain rule to calculate the adjoint of the affine Riemannian exponential Exp_P(A) = P^{1/2} Exp(P^{-1/2} A P^{-1/2}) P^{1/2}.
}
\details{
Calculates the adjoint of d_A Exp_P as

(d_A Exp_P)^\dagger (V) = P^{1/2} ((d_{P^{-1/2} A P^{-1/2}} Exp)^\dagger(P^{-1/2} V P^{-1/2}) P^{1/2}

This has the property that

< d_AExp_P(W), V >\emph{{Exp_P(A)} = < W, (d_A Exp_P)^\dagger(V) >}{P}

For all symmetric matrices W and V.
}
