% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine_differentials.R
\name{exponential_derivative}
\alias{exponential_derivative}
\title{exponential_derivative}
\usage{
exponential_derivative(A, V, t = 1, symm = TRUE, tol = 1e-08)
}
\arguments{
\item{A}{Matrix: The evaluation point for the matrix derivative}

\item{V}{Matrix: The direction in which to take the directional derivative}

\item{t}{Numeric: The time at which to take the derivative}

\item{symm}{Logical: Is A a symmetric matrix or not}

\item{tol}{Numeric: Tolerance at which we determine if eigenvalues are "close"}
}
\value{
A matrix that is the directional derivative at A in the direction V
}
\description{
Calculate the exponential derivative for a semi-simple matrix as in Najfeld and Havel (1995)
}
\details{
Calculates the directional derivative of Exp(tA) in the direction V.  Specifically, it calculates

D_v(t, A) = U (\bar{V} \odot \Theta(t)) U^{-1}

where \bar{V} = U^{-1} V U, \odot represents the Hadamard product, and \Theta(t) is the output of get_theta_matrix.
}
\examples{

set.seed(2021)
A <- matrix(rnorm(3 * 3), 3, 3)
A <- (A + t(A))/2

V <- matrix(rnorm(3 * 3), 3, 3)
V <- (V + t(V))/2

# Make sure exponential derivative matches finite difference

h <- 1e-8
(expm::expm(A + h*V) - expm::expm(A))/h
exponential_derivative(A, V)

}
