% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine_differentials.R
\name{exponential_derivative_general}
\alias{exponential_derivative_general}
\title{exponential_derivative_general}
\usage{
exponential_derivative_general(
  P,
  A,
  V,
  t = 1,
  symm = TRUE,
  P.sq = NULL,
  tol = 1e-08
)
}
\arguments{
\item{P}{Matrix: a symmetric positive definite matrix}

\item{A}{Matrix: The evaluation point in the tangent space at P}

\item{V}{Matrix: The direction in the tangent space at P to take the derivative in}

\item{t}{Numeric: The time point at which to take the derivative.}

\item{symm}{Logical: Whether the matrix A is symmetric or not}

\item{P.sq}{Matrix: Optional square root of P to speed up calculation time}

\item{tol}{Numeric: Tolerance at which we determine if eigenvalues are "close"}
}
\value{
The directional derivative of Exp_P(tA) in the direction V
}
\description{
Apply the chain rule to calculate the derivative of the affine exponential map
}
\examples{

set.seed(2021)
P <- matrix(rnorm(3 * 3), 3, 3)
P <- (P + t(P))/2
P <- expm::expm(P) # Make sure P is positive definite

A <- matrix(rnorm(3 * 3), 3, 3)
A <- (A + t(A))/2

V <- matrix(rnorm(3 * 3), 3, 3)
V <- (V + t(V))/2

# Make sure exponential derivative matches finite difference

h <- 1e-8
(affine_map(P, A + h*V) - affine_map(P, A))/h
exponential_derivative_general(P, A, V)

}
