% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_functions.R
\name{fitted.riemannian_pls}
\alias{fitted.riemannian_pls}
\title{fitted.riemannian_pls
Returns fitted values from riemannian pls model}
\usage{
\method{fitted}{riemannian_pls}(M, ...)
}
\arguments{
\item{...}{}
}
\value{
TBA
}
\description{
fitted.riemannian_pls
Returns fitted values from riemannian pls model
}
\examples{
TBA
}
