% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_simulation_functions.R
\name{generate_signal_matrix}
\alias{generate_signal_matrix}
\title{Generate a signal matrix}
\usage{
generate_signal_matrix(
  rho,
  block_size,
  method = "constant",
  rho_max = NULL,
  rho_min = NULL,
  vary = FALSE,
  se = 0.05,
  seed = NULL
)
}
\arguments{
\item{rho}{Numeric. The average strength of the signal to implant.  Should be between -1 and 1.}

\item{block_size}{Numeric. The size of the matrix to implant.  An integer greater than 0.}

\item{method}{Character. What method of correlation structure should be used.  The current options are
"constant", "ar1", or "hub".}

\item{rho_max}{Numeric. Used only for method = "hub".  The largest signal strength we want to measure.
Should be between -1 and 1.}

\item{rho_min}{Numeric. Used only for method = "hub".  The smallest signal strength we want to measure.
Should be between -1 and rho_max.}

\item{vary}{Logical. Do we add variation to the average signal strength.
This is used for more subject-to-subject variability.}

\item{se}{Numeric.  The standard error to be used if vary = TRUE.}

\item{seed}{Numeric.  If a value is given, this will be used as the seed for the signal generation.}
}
\value{
A matrix of size block_size x block_size used to implant signal into an FC matrix
}
\description{
Generate a signal matrix
}
\details{
TO BE FILLED OUT
}
\examples{
TO BE FILLED OUT
}
