% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine_differentials.R
\name{get_omega_matrix}
\alias{get_omega_matrix}
\title{get_omega_matrix}
\usage{
get_omega_matrix(D, tol = 1e-08)
}
\arguments{
\item{D}{\itemize{
\item Vector - the vector containing your eigenvalues.
}}

\item{tol}{Numeric - The tolerance to decide whether eigenvalues are "close."}
}
\value{
A square matrix Omega used to calculate the derivative of the logarithm
}
\description{
Calculate the Omega matrix used to get the derivative of the matrix logarithm
}
\details{
Calculates the square matrix Omega such that

omega_{ij} = omega_{ji} = (log(lambda_i) - log(lambda_j))/(lambda_i - lambda_j) if lambda_i \neq lambda_j and 1/lambda_i else.
}
