% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_short_functions.R
\name{get_sqrt}
\alias{get_sqrt}
\title{get_sqrt}
\usage{
get_sqrt(
  A,
  symmetric = isSymmetric(A, tol = sqrt(.Machine$double.eps)),
  method = c("custom", "expm")
)
}
\arguments{
\item{A}{\itemize{
\item A square matrix to find the square root of.
}}

\item{symmetric}{\itemize{
\item Logical: Is the matrix A symmetric or not.  Defaults to isSymmetric(A)
}}

\item{method}{\itemize{
\item Character: Either custom or expm, to choose how you want to calculate the square root.
}}
}
\value{
A matrix B such that B^2 = A (or B^T B = A in the non-symmetric case).
}
\description{
This function will obtain the square root of a matrix for you.  Currently two choices are available: The symmetric
sqaure root or Cholesky square root.  This is not optomised for large or poorly behaved matrices.
}
\examples{

## Symmetric square root
A <- matrix(
c(
  1, 0.5,
  0.5, 1)
 ),
 2, 2
)

get_sqrt(A)

## Non-symmetric square root

get_sqrt(A, symmetric = FALSE)


}
