% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine_differentials.R
\name{get_theta_matrix}
\alias{get_theta_matrix}
\title{get_theta_matrix}
\usage{
get_theta_matrix(D, t = 1, tol = 1e-08)
}
\arguments{
\item{D}{Vector - the vector containing your eigenvalues.}

\item{t}{Numeric - The time component for the matrix derivative.}

\item{tol}{Numeric - The tolerance to decide whether eigenvalues are "close"}
}
\value{
The Theta matrix used to calculate the matrix derivative.
}
\description{
Calcualte the Theta matrix used to get the derivative of the matrix exponential.
}
\details{
Returns the matrix \Theta such that

\theta_{ij} = \sinh(t(\lambda_i - \lambda_j)/2)/(t(\lambda_i - \lambda_j)/2)

if \lambda_i \neq \lambda_j and 1 else.
}
