% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_short_functions.R
\name{get_trace}
\alias{get_trace}
\title{get_trace}
\usage{
get_trace(A)
}
\arguments{
\item{A}{\itemize{
\item A square matrix
}}
}
\value{
The trace of A
}
\description{
Calculate the trace if A
}
