% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_pls_affine.R
\name{gradient_direction_affine}
\alias{gradient_direction_affine}
\title{Gradient of direction, affine variant
Calculates a component of the gradient of the weights/loading vectors for the affine-invariant geometry.}
\usage{
gradient_direction_affine(mu, P, X, t = 1, ...)
}
\arguments{
\item{mu}{Matrix: Positive definite symmetric matrix, mean of the X.}

\item{P}{Matrix: Current estimate of direction vector.}

\item{X}{Matrix:  Data matrix, i.e correlation matrix}

\item{t}{Numeric: Current estimate of projection coefficient}

\item{...}{\itemize{
\item check: logical, should inputs be checked.
}}
}
\value{
A matrix that is the portion of gradient of P attributed to X.
}
\description{
Gradient of direction, affine variant
Calculates a component of the gradient of the weights/loading vectors for the affine-invariant geometry.
}
\examples{

TBA
}
