% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_pls_affine.R
\name{gradient_direction_affine_full}
\alias{gradient_direction_affine_full}
\title{Calculate the gradient of the direction, affine variant.}
\usage{
gradient_direction_affine_full(mu, P, X, t, ...)
}
\arguments{
\item{mu}{Matrix: Mean of the X.}

\item{P}{Matrix: Current estimate of direction matrix.}

\item{X}{List: List of the X matrices.}

\item{t}{Numeric: Numeric vector of the current t values.}

\item{...}{Optional:
- mc.cores: Integer, to run in parallel, how many cores.
- mu.sq: Matrix, the square root of mu to speed up computation.
- check: Logical, should we check the variable?}
}
\value{
The full gradient of P at X and t.
}
\description{
Calculates the gradient for the direction (weight or loading) in the affine-invariant geometry.
}
\examples{

TBA
}
