% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_pls_affine.R
\name{gradient_score_affine}
\alias{gradient_score_affine}
\title{Gradient of the score, affine variant}
\usage{
gradient_score_affine(mu, W, X, t = 1, ...)
}
\arguments{
\item{mu}{Matrix: Positive definite symmetric matrix, mean of the X.}

\item{W}{Matrix: Current estimate of weights matrix W.}

\item{X}{Matrix: Data matrix, i.e. correlation matrix.}

\item{t}{Numeric: The current estimate of the projection coefficient t.}

\item{...}{\itemize{
\item check: logical, should inputs be checked.
}}
}
\value{
A value that is the gradient for t at X and W.
}
\description{
Calculates the gradient for calculating the score when dealing with the affine-invariant geometry
}
\examples{

TBA
}
