% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_simulation_functions.R
\name{implant_signal}
\alias{implant_signal}
\title{Implant signal matrix}
\usage{
implant_signal(A, B, rois = 4:7, r = NULL, chol = FALSE)
}
\arguments{
\item{A}{Matrix. The matrix we want to implant signal in.  The dimension of A must be larger than the dimension of B.}

\item{B}{Matrix. The signal we wish to implant}

\item{rois}{Integer vector.  The ids of the ROIs you want to plant a signal in.}

\item{r}{Numeric. Value between 0 and 1 which determines how much signal to implant.}
}
\value{
Matrix of the same dimension as A with B implanted into the top left block.
}
\description{
Implant signal matrix
}
\details{
Given

$$
A = \begin{bmatrix} A_{11} & A_{12} \\ A_{21} & A_{22} \end{bmatrix}
$$

Let

$$
C = \begin{bmatrix} B_c' ((A_{11})_c^{-1})' & 0 \\ 0 & I \end{bmatrix}
$$

where the subscript c represents the Cholesky square root.  The resulting matrix given by

$$
\tilde{A} = C A C'
$$

where the prime indicates matrix transpose, is such that $$\tilde{A}_{11} = B\, .$$

NOTE: At the moment, this function only implants signal into ascending order ROIS.
This can be further specialized to allow for more intrecate signals to be implanted.
}
\examples{
A <- diag(1, 10)
B <- matrix(5, nrow = 2, ncol = 2)

A.tilde <- implant_signal(A, B, rois = c(1, 5))
}
