% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_pls.R
\name{initiate_u}
\alias{initiate_u}
\title{Initiate u}
\usage{
initiate_u(Y, muY, ...)
}
\arguments{
\item{Y}{List: List of correlation matrices.  The data.}

\item{muY}{Matrix: Frechet mean of Y.}

\item{...}{Optional arguments:
- check: logical, should inputs be checked.
- muY.sq: Matrix, square root of muY to speed up computation}
}
\value{
A vector u of initial estimates of the scores
}
\description{
Initiate the score vector u in the NIPALS algorithm
}
\details{
First, each Y_i is mapped to the tangent space of muY via Log_{muY}.  Then these are vectorised
using the vec operation.  Finally, the first column of the resulting matrix is returned.
}
\examples{

TBA
}
