% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_short_functions.R
\name{is_positive_definite}
\alias{is_positive_definite}
\title{is_positive_definite}
\usage{
is_positive_definite(P, tol = 1e-10)
}
\arguments{
\item{P}{\itemize{
\item A matrix
}}

\item{tol}{\itemize{
\item A tolerance level
}}
}
\value{
Logical: TRUE if the matrix is positive definite, FALSE otherwise
}
\description{
Check to see if a matrix is positive definite.
}
\details{
Uses the eigen-value decomposition to determine is a matrix is positive definite up to a certain tolerance.
}
