% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_pls.R
\name{linearise_data}
\alias{linearise_data}
\title{Linearise the data to apply the tangent space method.}
\usage{
linearise_data(X, mu, ...)
}
\arguments{
\item{X}{List: List of SPD matrices}

\item{mu}{Matrix: Mean of X}

\item{...}{Optional list
\itemize{
\item mc.cores: Numeric.  Number of cores to run on.
\item mu.sq: Matrix.  Square root of muX.
}}
}
\value{
A matrix that is the tangent-space vectorisation of the X-data
}
\description{
Linearise the data to apply the tangent space method.
}
\examples{
TBA
}
