% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine_differentials.R
\name{log_adjoint}
\alias{log_adjoint}
\title{log_adjoint}
\usage{
log_adjoint(A, V, symm = TRUE, A.sq = NULL, tol = 1e-08)
}
\arguments{
\item{A}{Matrix: Evaluation point.  Must be symmetric positive definite.}

\item{V}{Matrix: Directional point.  Should be symmetric.}

\item{symm}{Logical: Is V symmetric.}

\item{A.sq}{Matrix: Optional square root of A to speed up calculations.}

\item{tol}{Numeric: Tolerance to which eigenvalues are "the same."}
}
\value{
A matrix of the same dimension as A which is the adjoint of the linear map d_ALog.
}
\description{
Calculates the adjoint of the linear map d_ALog, where Log is the matrix logarithm.
}
\details{
Calculates

(d_A Log)^\dagger(V) = A^{1/2} (d_{A}Log(A^{1/2} V A^{1/2}))A^{1/2}

Such that

< d_A Log(V), W >_I = < V, (d_A Log)^\dagger(W) >_A
}
