% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine_differentials.R
\name{log_adjoint_general}
\alias{log_adjoint_general}
\title{log_adjoint_general}
\usage{
log_adjoint_general(P, A, V, P.sq = NULL, symm = TRUE, tol = 1e-08)
}
\arguments{
\item{P}{Matrix: The base point.  Must be symmetric positive definite.}

\item{A}{Matrix: The evaluation point.  Must be symmetric positive definite.}

\item{V}{Matrix: The directional point.  Should be symmetric.}

\item{P.sq}{Matrix: Optional square root of P to speed up calculation.}

\item{symm}{Logical: Is V symmetric?}

\item{tol}{Numeric: Tolerance to which eigenvalues are "the same."}
}
\value{
A matrix of the same dimension as A which is the adjoint of the linear map d_ALog_P.
}
\description{
log_adjoint_general
}
\details{
Calculates the adjoint of d_A Log_P as

(d_A Log_P)^\dagger (V) = P^{1/2} ((d_{P^{-1/2} A P^{-1/2}} Log)^\dagger(P^{-1/2} V P^{-1/2}) P^{1/2}

This has the property that

< d_A Log_P(W), V >\emph{{P} = < W, (d_A Log_P)^\dagger(V) >}{A}

For all symmetric matrices W and V.
}
