% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine_differentials.R
\name{log_derivative}
\alias{log_derivative}
\title{log_derivative}
\usage{
log_derivative(A, V, symm = TRUE, tol = 1e-08)
}
\arguments{
\item{A}{Matrix: Evaluation point.  Must be positive definite.}

\item{V}{Matrix: Directional point.  Should be symmetric.}

\item{symm}{Logical: Is V a symmetric matrix.}

\item{tol}{Numeric: Tolerance at which we determine if eigenvalues are "close"}
}
\value{
A matrix of the same dimension as A that is the directional derivative of Log(A), the matrix logarithm, in the direction V.
}
\description{
Calculates the differential of the matrix logarithm at A in the direction V.
}
\details{
The differential d_A Log is found by using the fact that d_{Log(A)} Exp \circ d_A Log (V) = V.  This leverages the fact that
Log is the inverse of Exp, hence the differential of the composition is the identity mapping.
}
