% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine_differentials.R
\name{log_derivative_general}
\alias{log_derivative_general}
\title{log_derivative_general}
\usage{
log_derivative_general(P, A, V, symm = TRUE, tol = 1e-08)
}
\arguments{
\item{P}{Matrix: Base point.  Must be symmetric positive definite.}

\item{A}{Matrix: Evaluation point.  Must be symmetric positive definite.}

\item{V}{Matrix: Directional point.  Should be symmetric.}

\item{symm}{Logical: Is V a symmetric matrix.}

\item{tol}{Numeric: Tolerance at which we determine if eigenvalues are "close"}
}
\value{
A matrix of the same dimension as A that is the directional derivative of Log_P(A), the matrix logarithm, in the direction V.
}
\description{
Applies the chain rule to calculate the derivative of Log_P(A) in the direction V.
}
\details{
Uses the chain rule to calculate that

d_ALog_P(V) = P^{1/2} (d_{P^{-1/2} A P^{-1/2}} Log (P^{-1/2} V P^{-1/2})) P^{1/2}.
}
