% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_short_functions.R
\name{multiply_and_reduce}
\alias{multiply_and_reduce}
\title{multiply_and_reduce}
\usage{
multiply_and_reduce(t, W)
}
\arguments{
\item{t}{Numeric: a numeric vector of coefficient values}

\item{W}{List: A list of numeric matrices}
}
\value{
A matrix S the same size as the matrices in W.
}
\description{
Take a set of numeric values and a list of matrices, multiply them, and add them up.
}
\details{
Performs the calculation

S = sum( t_i * W_i )
}
\examples{

t <- 1:3

W <- map(1:3, ~diag(1, 3))

multiply_and_reduce(t, W)

}
