% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_functions.R
\name{predict.riemannian_pls}
\alias{predict.riemannian_pls}
\title{predict.riemannian_pls}
\usage{
\method{predict}{riemannian_pls}(object, newdata, tol = 1e-05, max.iter = 50, tau = 1/2, ...)
}
\arguments{
\item{object}{\itemize{
\item the Riemannian PLS model
}}

\item{newdata}{\itemize{
\item New X data to predict on
}}

\item{tol}{\itemize{
\item Tolerance at which to calc t-values
}}

\item{max.iter}{\itemize{
\item Max iter at which to calc t-values
}}

\item{tau}{\itemize{
\item Step size
}}

\item{...}{}
}
\value{
TBA
}
\description{
predict.riemannian_pls
}
\examples{
TBA
}
