% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_pls.R
\name{riemannian_pls}
\alias{riemannian_pls}
\title{Fit the Riemannian PLS model}
\usage{
riemannian_pls(
  X,
  Y,
  L = 1,
  tol = 1e-07,
  max.iter = 50,
  muX = NULL,
  muY = NULL,
  tau = 1/2,
  ...
)
}
\arguments{
\item{X}{List or matrix: Predictor data.  List for affine, matrix for euclidean}

\item{Y}{List or matrix: Response data.  List for affine, matrix for euclidean}

\item{L}{Numeric: Number of PLS components to calculate}

\item{tol}{Numeric: The tolerance/precision level to work at}

\item{max.iter}{Numeric: The maximum number of iterations in the while loops}

\item{muX}{Matrix:  If X is affine, the frechet mean of X}

\item{muY}{Matrix:  If Y is affine, the frechet mean of Y}

\item{tau}{Numeric: step size in gradient descent}

\item{...}{Optional list
- type_x, type_y: String. The data type.  Can be "euclidean" or "affine".  Defualts to affine.
- method: String.  Are we performing the tangent-space method.  Only applicable if data is affine.  Only available choice is method = "tangent"
- centre: Logical.  If X, Y are Euclidean, do we need to centre?  Suggested no if method = "tangent"
- muX.sq, muY.sq: Matrix.  The square roots of muX and muY.
- mc.cores: Numeric.  The number of cores to run on.
- check: Logical.  Should we check the input data types.
- u, t, W, C: Initial estimates of u, t, W, and C.
- verbose: Logical.  Should we print convergence output.
- Other options to be passed to scores, loadings, regression, and deflate functions.}
}
\value{
A list containing:
\itemize{
\item scoresX: List of scores for X at each PLS step
\item scoresY: List of scores for Y at each PLS step
\item weightsX: List of weights for X at each PLS step
\item weightsY: List of weights for Y at each PLS step
\item loadingsX: List of loadings for X at each PLS step
\item reg_steps: List containing the inner relationship data at each step
\item L: Numeric: How many PLS directions were calculated.
\item X: List: contains the deflated X-data, position l + 1 contains the data used to calculate PLS direction l
\item Y: List: contains the deflated Y-data, position l + 1 contains the data used to calculate PLS direction l
}
}
\description{
Fit the Riemannian PLS model
}
\examples{
TBA
}
