% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_functions.R
\name{rmse.riemannian_pls}
\alias{rmse.riemannian_pls}
\title{rmse.riemannian_pls
Calculates RMSE from Riemannian PLS model}
\usage{
rmse.riemannian_pls(truth, est, ...)
}
\arguments{
\item{truth}{\itemize{
\item True data
}}

\item{est}{\itemize{
\item Estimated data
}}

\item{...}{}
}
\value{
TBA
}
\description{
rmse.riemannian_pls
Calculates RMSE from Riemannian PLS model
}
\examples{
TBA
}
