% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_functions.R
\name{rsq.riemannian_pls}
\alias{rsq.riemannian_pls}
\title{rsq.riemannian_pls
Returns R^2 value for Riemannian PLS model}
\usage{
rsq.riemannian_pls(
  truth,
  est,
  muY = NULL,
  m = NULL,
  ignore = TRUE,
  test = FALSE,
  ...
)
}
\arguments{
\item{truth}{\itemize{
\item True data
}}

\item{est}{\itemize{
\item Estimated data
}}

\item{muY}{\itemize{
\item True muY
}}

\item{m}{\itemize{
\item column mean if euclidean
}}

\item{ignore}{\itemize{
\item Logical, should I calculate means anyway
}}

\item{...}{}
}
\value{
TBA
}
\description{
rsq.riemannian_pls
Returns R^2 value for Riemannian PLS model
}
\examples{
TBA
}
