% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_pls.R
\name{scores_and_weights}
\alias{scores_and_weights}
\title{Calculate the scores and weights for Riemannian PLS}
\usage{
scores_and_weights(
  X,
  Y,
  muX = NULL,
  muY = NULL,
  max.iter = 50,
  tol = 1e-06,
  tau = 1/2,
  ...
)
}
\arguments{
\item{X}{List or matrix: Predictor data.  List for affine, matrix for euclidean}

\item{Y}{List or matrix: Response data.  List for affine, matrix for euclidean}

\item{muX}{Matrix:  If X is affine, the frechet mean of X}

\item{muY}{Matrix:  If Y is affine, the frechet mean of Y}

\item{max.iter}{Numeric: The maximum number of iterations in the while loops}

\item{tol}{Numeric: The tolerance/precision level to work at}

\item{tau}{Numeric: Step size in the gradient descent.}

\item{...}{Optional list:
- type_x, type_y: String. The data type.  Can be "euclidean" or "affine".  Defualts to affine.
- muX.sq, muY.sq: Matrix.  The square roots of muX and muY.
- mc.cores: Numeric.  The number of cores to run on.
- check: Logical.  Should we check the input data types.
- u, t, W, C: Initial estimates of u, t, W, and C.
- verbose: Logical.  Should we print convergence output.}
}
\value{
A list of values containing the estimates of u, t, W and C.
}
\description{
Calculate the scores and weights for Riemannian PLS
}
\examples{
TBA
}
