% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_simulation_functions.R
\name{simulate_subject}
\alias{simulate_subject}
\title{Simulate a subject}
\usage{
simulate_subject(
  rho,
  block_size,
  group,
  method = "constant",
  test = FALSE,
  low_rank = FALSE,
  ...
)
}
\arguments{
\item{rho}{Numeric. The average signal strength to be implanted. Should be between -1 and 1.}

\item{block_size}{Numeric. The size of the matrix to implant.  An integer greater than 0.}

\item{group}{Character. Is the subject a "patient" or "control."
Patients will have an implanted signal, controls will not.}

\item{method}{Character. What method of correlation structure should be used.  The current options are
"constant", "ar1", or "hub".}

\item{test}{Logical. This is for testing purposes to check simulations against truth.}

\item{low_rank}{Logical.  Should some matrices be converted to low rank to test regularisation?}

\item{...}{Other inputs for generate_signal_matrix.}
}
\value{
A single simulated functional connectivity matrix, with signal implanted if they are a patient.
}
\description{
Simulate a subject
}
\details{
As a base, the subject will be randomly assigned a functional connectivity matrix from
a set of true data (our method used the COBRE data and the MSDL atlas).  A signal matrix will then
be created using the generate_signal_matrix function using the variables specified in the function
input.  This signal will be implanted using the implant_signal function.  Finally, the resulting
matrix with the implanted signal will be used as the scale matrix in a Wishart distribution
to act as the simulated data.
}
\examples{
TO BE FILLED OUT
}
