% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine_functions.R
\name{translate_to_identity}
\alias{translate_to_identity}
\title{translate_to_identity}
\usage{
translate_to_identity(P, W, chol = FALSE, reverse = FALSE, P.sq = NULL)
}
\arguments{
\item{P}{Matrix: The point on the manifold}

\item{W}{Matrix: The point in TpM}

\item{chol}{Logical: Do you want ot use the Cholesky square root?}

\item{reverse}{Logical: Do you want to translate from the identity to P}

\item{P.sq}{Matrix: Optional. The square root of P, to help speed calculation.}
}
\value{
The translated matrix at the identity.
}
\description{
A function to translate a point in the tangent space of P to the tangent space of the identity
by the affine group action.
}
\details{
Performs the translation

\if{html}{\out{<div class="sourceCode">}}\preformatted{P^\{-1/2\} W P^\{-1/2\}
}\if{html}{\out{</div>}}

if reverse is FALSE to move W to the tangent space of the identity.  If reverse is TRUE, then
applies

\if{html}{\out{<div class="sourceCode">}}\preformatted{P^\{1/2\} W P^\{1/2\}
}\if{html}{\out{</div>}}

To move W from the tangent space at the identity to the tangent space at P.  Although this is
intended for use on tangent vectors W, this can be applied to any symmetric matrix.
}
\examples{

## Get some matrices
## We will look at the first two subjects in the COBRE dataset
P <- cobre$cors[[1]]
W <- cobre$cors[[2]]

## Translate W to a tangent vector at the identity:

W_trans <- translate_to_identity(P, W)

## Translate back to W

W_back <- translate_to_identity(P, W_trans, reverse = TRUE)

## Check it is the reverse (up to numerical error)
identical(zapsmall(W_back), zapsmall(W))

}
